# optimize encode