Avoid branches in the main encode loop.


```
benchmark                     old ns/op     new ns/op     delta
BenchmarkEncodeToString-4     19981         15283         -23.51%
BenchmarkDecodeString-4       46700         48240         +3.30%

benchmark                     old MB/s     new MB/s     speedup
BenchmarkEncodeToString-4     409.98       536.02       1.31x
BenchmarkDecodeString-4       280.77       271.80       0.97x
```